Document, restucture and cleanup Gateway
