+++ apple/Sources/Sargon/Antennas/Network/NetworkAntenna+URLSession.swift
@@ -6,9 +6,9 @@ extension URLRequest {
 		var request = URLRequest(url: sargon.url)
 		switch sargon.method {
 		case .post:
-			request.httpMethod = "POST" // FIXME: embed in sargon
+			request.httpMethod = "POST"  // FIXME: embed in sargon
 		}
-		
+
 		request.httpBody = sargon.body
 		request.allHTTPHeaderFields = sargon.headers
 		self = request
@@ -18,7 +18,8 @@ extension URLRequest {
 extension NetworkResponse {
 	init(response: (Data, URLResponse)) throws {
 		guard let httpURLResponse = response.1 as? HTTPURLResponse else {
-			throw SargonError.NetworkRequestGenericFailure(underlying: "Failed to cast to HTTPURLResponse")
+			throw SargonError.NetworkRequestGenericFailure(
+				underlying: "Failed to cast to HTTPURLResponse")
 		}
 		self.init(
 			statusCode: UInt16(httpURLResponse.statusCode),
@@ -28,8 +29,8 @@ extension NetworkResponse {
 }
 
 extension URLSession: NetworkAntenna {
-	
-	public func makeRequest(
+
+	public func executeNetworkRequest(
 		request sargonRequest: NetworkRequest
 	) async throws -> NetworkResponse {
 		let request = URLRequest(sargon: sargonRequest)
@@ -37,7 +38,8 @@ extension URLSession: NetworkAntenna {
 		do {
 			response = try await data(for: request)
 		} catch {
-			throw SargonError.NetworkRequestGenericFailure(underlying: String(describing: error))
+			throw SargonError.NetworkRequestGenericFailure(
+				underlying: String(describing: error))
 		}
 		return try NetworkResponse(response: response)
 	}
