+++ src/gateway_api/client/mod.rs
@@ -1,0 +1,5 @@
+mod gateway_client;
+mod gateway_client_dispatch_request;
+
+pub use gateway_client::*;
+pub use gateway_client_dispatch_request::*;
