+++ src/gateway_api/mod.rs
@@ -1,5 +1,11 @@
-mod gateway_client;
+mod client;
+mod endpoints;
+mod methods;
 mod models;
+mod network_antenna;
 
-pub use gateway_client::*;
+pub use client::*;
+pub use endpoints::*;
+pub use methods::*;
 pub use models::*;
+pub use network_antenna::*;
