+++ src/gateway_api/network_antenna/network_method.rs
@@ -1,0 +1,6 @@
+use crate::prelude::*;
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq, uniffi::Enum)]
+pub enum NetworkMethod {
+    Post,
+}
