+++ src/gateway_api/network_antenna/network_request.rs
@@ -1,0 +1,10 @@
+use crate::prelude::*;
+
+#[derive(Clone, Debug, PartialEq, Eq, uniffi::Record)]
+pub struct NetworkRequest {
+    pub url: Url,
+    pub method: NetworkMethod,
+    pub headers: HashMap<String, String>,
+
+    pub body: BagOfBytes,
+}
